Add function to output CSV and refactor code for readability.
Move all functions related to CSV file management (such as generating input and output
locations, etc.) to their own translation unit with associated header. This improves
readability of the main parser.cpp file.

Create a function to output a CSV based on the processed rows. Base assumption is the
filename should be the same as the input file with '_Processed' appended to it, but
possible that this will change in the future.